+++ dev.yml
@@ -3,4 +3,7 @@ version: '3'
 services:
   wolf:
     ports:
-      - "8000:80"
+      - '8000:80'
+  mongo:
+    ports:
+      - '27017:27017'
