+++ docker-compose.yml
@@ -5,8 +5,8 @@ services:
     build: .
     image: parceldock
     ports:
-      - "1234:1234"
-      - "1235:1235"
+      - '1234:1234'
+      - '1235:1235'
     volumes:
       - .:/app/
       - ./dist/:/app/dist/
@@ -14,7 +14,7 @@ services:
     image: wolf
     build: wolf/
     volumes:
-      - ./wolf:/wolf
+      - ./leo:/wolf
       - ./dist:/wolf/dist
       - ./pdf:/wolf/pdf
     depends_on:
