+++ leo/api.py
@@ -1,18 +1,20 @@
-from fastapi import FastAPI, Path, Body
-from starlette.staticfiles import StaticFiles
-from uvicorn import run
-from os import environ
-from setlist import setlist
+from os import path
+
+from fastapi import Body, FastAPI, Path
 from pymongo import MongoClient
+from setlist import setlist
+from starlette.staticfiles import StaticFiles
 
 client = MongoClient('mongodb://mongo:27017', connect=False)
 db = client.leo
 
 app = FastAPI()
 
+
 @app.get('/setlist')
 def getSetlist():
     return setlist('pdf')
+
 
 @app.get('/songs')
 def getSongs(s: str = ''):
@@ -26,11 +28,17 @@ def getAnnotations(song: str = Path(..., title='name of song')):
     if results:
         return results['annotations']
 
+
 @app.post('/annotations/{song}')
-def postAnnotations(*, annotations=Body(...), song: str = Path(..., title='name of song')):
+def postAnnotations(
+    *, annotations=Body(...), song: str = Path(..., title='name of song')
+):
     print('saved annotations!', song, annotations)
-    res = db.annotations.update({ 'song': song }, { '$set': {'annotations': annotations} }, upsert=True)
+    res = db.annotations.update(
+        {'song': song}, {'$set': {'annotations': annotations}}, upsert=True
+    )
     print('saved! ', res)
+
 
 app.mount("/pdf", StaticFiles(directory='pdf'))
 app.mount("/", StaticFiles(directory='dist', html=True), name="static")
