+++ leo/main.py
@@ -1,4 +1,6 @@
 from api import app
 
 if __name__ == '__main__':
+    from uvicorn import run
+
     run(app, host='0.0.0.0', port=80)
