+++ leo/setlist.py
@@ -1,5 +1,6 @@
 import os
 
+
 def setlist(path):
     setlist = []
     for dpath, subdirs, files in os.walk(path):
