+++ pyproject.toml
@@ -1,0 +1,25 @@
+[project]
+name = "leo"
+version = "0.1.0"
+description = "A virtual conducting webapp"
+authors = [
+    {name = "Quaternion Media",email = "leo@quaternion.media"},
+]
+dependencies = [
+    "fastapi>=0.103.1",
+    "pymongo>=4.5.0",
+    "pyRealParser>=0.1.0",
+    "aiofiles>=23.2.1",
+]
+requires-python = ">=3.11"
+readme = "README.md"
+license = {text = "MIT"}
+
+[build-system]
+requires = ["pdm-backend"]
+build-backend = "pdm.backend"
+
+[tool.pdm.dev-dependencies]
+dev = [
+    "uvicorn>=0.23.2",
+]
